[All] Properly spaced '*' and gave proper types

+++ KeyboardSupport/RAKeyboardWindow.h
@@ -5,8 +5,8 @@
 	RARemoteKeyboardView *kbView;
 }
 
-- (void)setupForKeyboardAndShow:(NSString*)identifier;
+- (void)setupForKeyboardAndShow:(NSString *)identifier;
 - (void)removeKeyboard;
 
-- (unsigned int)contextId;
+- (NSUInteger)contextId;
 @end
